separate api from unit actions
